convert non-extension mech-and-cow to *.cow format
Convert the non-extension mech-and-cow file to the *.cow
format, taking some hints from the MechAndCow.pm source
for the location of the $thoughts line, and from the
default.cow file for the $eyes and $tongue.

The mech-and-cow.cow file can probably be removed in the
future once the *.pm format has been implemented.

The $thoughts line to the bubble ends quite far away from
the left edge of the screen (column 18) which looks weird
for narrow messages, but other cows end even further to
the right (e.g dragon-and-cow.cow in column 23), so this
should be ok.