Make source tree cowthink.1 a symlink to cowsay.1
As the source tree is already using symlinks for the cowthink script,
we can assume working symlinks and therefore can use them for the
cowthink.1 man page as well.

This has the advantage that when the man page is updated, no special
treatment is required for the "man1/cowthink.1" reference inside
the cowthink.1 file produced by a2x.

Note that due to an apparent limitation of the git tool, an earlier
commit must remove the regular file "cowthink.1" before this commit
can add a symlink called "cowthink.1".