Use man page "alias" instead of symlink for cowthink.1
This alias file is a groff macro (I think) that reads in the referenced file. I'm going to stick with this way instead of the special symlink in the installation location, because it's more in line with what the Asciidoctor toolchain does, and reduces differences between the in-repo and installed-to-prefix versions of cowsay.