Reliable symlink creation through prior deletion
As you cannot rely on "ln -s foo bar" to do the right thing if
the something of the same name already exists, it is prudent to
remove the name before creating it as a symlink.

And "ln -sf" has its own issues. So... "rm -f ... && ln -s" it is.