Move man source files to separate man-src dir, to avoid "ignoring bogus filename" warning
On Fedora (but not Ubuntu or macOS, it seems), if you put the in-repo man/ dir on the manpath and do `man cowsay`, if the corresponding .adoc source file is there, it'll give you a warning. This breaks the man dir in to separate source and built dirs, and should get rid of that warning.

Example on Fedora 39:

```
janke@fed39-test-as:~/repos/cowsay$ PATH=$PWD/bin:$PATH
janke@fed39-test-as:~/repos/cowsay$ which cowsay
~/repos/cowsay/bin/cowsay
janke@fed39-test-as:~/repos/cowsay$ man -a -w cowsay
man: warning: /home/janke/repos/cowsay/man/man1/cowsay.1.adoc: ignoring bogus filename
/home/janke/repos/cowsay/man/man1/cowsay.1
janke@fed39-test-as:~/repos/cowsay$ man -a -w cowthink
/home/janke/repos/cowsay/man/man1/cowsay.1
janke@fed39-test-as:~/repos/cowsay$
```