Add `animalsay` invocation mode and `-H` thinking option
Invoking as `animalsay` will give you a random cow, like `cowsay -r`. This `animalsay` alias is functionality that Fedora's package customization was providing. Pulling it upstream here means packagers don't have to do a customization, and its random mode supports our $COWPATH and related functionality, so all cows are available to the random selection.

The new `-H` option activates thinking mode, just like invoking the program as `cowthink`. This is mostly for debugging purposes, and a bit for aesthetic "completeness".