Apply `use utf8;` to existing UTF8-encoded cowfiles
Somehow, activating UTF8 support on input & output streams and args messes these cows up. (I suspect it's because Perl is now doing UTF8 encoding on STDOUT, so it reads the source for these cowfiles as a single-byte encoding like Latin-1, misinterprets the characters, and then encodes the output as the correct UTF8 representation of those wrong characters.) Marking the source files as UTF-8 with the `use utf8;` pragma in the source protects against that.